backend/requirements.txt
# eSIM Myanmar Entertainment Server - Python Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
pymongo==4.6.1
motor==3.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment & Configuration
python-dotenv==1.0.0

# QR Code Generation
qrcode==7.4.2
Pillow==10.2.0

# HTTP Clients
requests==2.31.0
aiohttp==3.9.1
httpx==0.26.0

# WebSocket Support
websockets==12.0

# Utilities
python-dateutil==2.8.2

# Testing (Development)
pytest==7.4.4
pytest-asyncio==0.23.3
faker==22.0.0
httpx==0.26.0

# Monitoring (Optional)
prometheus-client==0.19.0